requirements.txt
python-telegram-bot[http2]==21.5
httpx==0.27.2
aiosqlite==0.20.0
python-dotenv==1.0.1